3.8*: Phonebook backend step8
Configure morgan so that it also shows the data sent in HTTP POST requests:
fullstack content
Note that logging data even in the console can be dangerous since it can contain sensitive data and may violate local privacy law (e.g. GDPR in EU) or business-standard. In this exercise, you don't have to worry about privacy issues, but in practice, try not to log any sensitive data.
This exercise can be quite challenging, even though the solution does not require a lot of code.
This exercise can be completed in a few different ways. One of the possible solutions utilizes these two techniques:
    creating new tokens
    JSON.stringify